% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudkms_functions.R
\docType{package}
\name{cloudkms_googleAuthR}
\alias{cloudkms_googleAuthR}
\alias{cloudkms_googleAuthR-package}
\title{Google Cloud Key Management Service (KMS) API
Manages encryption for your cloud services the same way you do on-premise. You can generate, use, rotate, and destroy AES256 encryption keys.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:32:30
filename: /Users/mark/dev/R/autoGoogleAPI/googlecloudkmsv1beta1.auto/R/cloudkms_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}
