% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudkms_objects.R
\name{CryptoKey}
\alias{CryptoKey}
\title{CryptoKey Object}
\usage{
CryptoKey(primary = NULL, name = NULL, purpose = NULL,
  nextRotationTime = NULL, createTime = NULL, rotationPeriod = NULL)
}
\arguments{
\item{primary}{Output only}

\item{name}{Output only}

\item{purpose}{The immutable purpose of this CryptoKey}

\item{nextRotationTime}{At next_rotation_time, the Key Management Service will automatically:}

\item{createTime}{Output only}

\item{rotationPeriod}{next_rotation_time will be advanced by this period when the service}
}
\value{
CryptoKey object
}
\description{
CryptoKey Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A CryptoKey represents a logical key that can be used for cryptographicoperations.A CryptoKey is made up of one or more versions, whichrepresent the actual key material used in cryptographic operations.
}
