% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudkms_objects.R
\name{CryptoKeyVersion}
\alias{CryptoKeyVersion}
\title{CryptoKeyVersion Object}
\usage{
CryptoKeyVersion(state = NULL, name = NULL, destroyEventTime = NULL,
  destroyTime = NULL, createTime = NULL)
}
\arguments{
\item{state}{The current state of the CryptoKeyVersion}

\item{name}{Output only}

\item{destroyEventTime}{Output only}

\item{destroyTime}{Output only}

\item{createTime}{Output only}
}
\value{
CryptoKeyVersion object
}
\description{
CryptoKeyVersion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A CryptoKeyVersion represents an individual cryptographic key, and theassociated key material.It can be used for cryptographic operations either directly, or via itsparent CryptoKey, in which case the server will choose the appropriateversion for the operation.
}
