% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudkms_objects.R
\name{DecryptRequest}
\alias{DecryptRequest}
\title{Google Cloud Key Management Service (KMS) API Objects 
Manages encryption for your cloud services the same way you do on-premise. You can generate, use, rotate, and destroy AES256 encryption keys.}
\usage{
DecryptRequest(ciphertext = NULL, additionalAuthenticatedData = NULL)
}
\arguments{
\item{ciphertext}{Required}

\item{additionalAuthenticatedData}{Optional data that must match the data originally supplied in}
}
\value{
DecryptRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:32:30
filename: /Users/mark/dev/R/autoGoogleAPI/googlecloudkmsv1beta1.auto/R/cloudkms_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
DecryptRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for KeyManagementService.Decrypt.
}
