% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudkms_objects.R
\name{DecryptResponse}
\alias{DecryptResponse}
\title{DecryptResponse Object}
\usage{
DecryptResponse(plaintext = NULL)
}
\arguments{
\item{plaintext}{The decrypted data originally supplied in EncryptRequest}
}
\value{
DecryptResponse object
}
\description{
DecryptResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for KeyManagementService.Decrypt.
}
