% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudkms_objects.R
\name{EncryptRequest}
\alias{EncryptRequest}
\title{EncryptRequest Object}
\usage{
EncryptRequest(additionalAuthenticatedData = NULL, plaintext = NULL)
}
\arguments{
\item{additionalAuthenticatedData}{Optional data that, if specified, must also be provided during decryption}

\item{plaintext}{Required}
}
\value{
EncryptRequest object
}
\description{
EncryptRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for KeyManagementService.Encrypt.
}
