% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudkms_objects.R
\name{EncryptResponse}
\alias{EncryptResponse}
\title{EncryptResponse Object}
\usage{
EncryptResponse(ciphertext = NULL, name = NULL)
}
\arguments{
\item{ciphertext}{The encrypted data}

\item{name}{The resource name of the CryptoKeyVersion used in encryption}
}
\value{
EncryptResponse object
}
\description{
EncryptResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for KeyManagementService.Encrypt.
}
