% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudkms_objects.R
\name{KeyRing}
\alias{KeyRing}
\title{KeyRing Object}
\usage{
KeyRing(createTime = NULL, name = NULL)
}
\arguments{
\item{createTime}{Output only}

\item{name}{Output only}
}
\value{
KeyRing object
}
\description{
KeyRing Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A KeyRing is a toplevel logical grouping of CryptoKeys.
}
