% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudkms_objects.R
\name{ListCryptoKeysResponse}
\alias{ListCryptoKeysResponse}
\title{ListCryptoKeysResponse Object}
\usage{
ListCryptoKeysResponse(nextPageToken = NULL, cryptoKeys = NULL,
  totalSize = NULL)
}
\arguments{
\item{nextPageToken}{A token to retrieve next page of results}

\item{cryptoKeys}{The list of CryptoKeys}

\item{totalSize}{The total number of CryptoKeys that matched the query}
}
\value{
ListCryptoKeysResponse object
}
\description{
ListCryptoKeysResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for KeyManagementService.ListCryptoKeys.
}
