% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudkms_objects.R
\name{ListCryptoKeyVersionsResponse}
\alias{ListCryptoKeyVersionsResponse}
\title{ListCryptoKeyVersionsResponse Object}
\usage{
ListCryptoKeyVersionsResponse(nextPageToken = NULL, totalSize = NULL,
  cryptoKeyVersions = NULL)
}
\arguments{
\item{nextPageToken}{A token to retrieve next page of results}

\item{totalSize}{The total number of CryptoKeyVersions that matched the}

\item{cryptoKeyVersions}{The list of CryptoKeyVersions}
}
\value{
ListCryptoKeyVersionsResponse object
}
\description{
ListCryptoKeyVersionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for KeyManagementService.ListCryptoKeyVersions.
}
