% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudkms_objects.R
\name{ListKeyRingsResponse}
\alias{ListKeyRingsResponse}
\title{ListKeyRingsResponse Object}
\usage{
ListKeyRingsResponse(keyRings = NULL, nextPageToken = NULL,
  totalSize = NULL)
}
\arguments{
\item{keyRings}{The list of KeyRings}

\item{nextPageToken}{A token to retrieve next page of results}

\item{totalSize}{The total number of KeyRings that matched the query}
}
\value{
ListKeyRingsResponse object
}
\description{
ListKeyRingsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for KeyManagementService.ListKeyRings.
}
