% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudkms_objects.R
\name{Location}
\alias{Location}
\title{Location Object}
\usage{
Location(Location.metadata = NULL, Location.labels = NULL, name = NULL,
  locationId = NULL, metadata = NULL, labels = NULL)
}
\arguments{
\item{Location.metadata}{The \link{Location.metadata} object or list of objects}

\item{Location.labels}{The \link{Location.labels} object or list of objects}

\item{name}{Resource name for the location, which may vary between implementations}

\item{locationId}{The canonical id for this location}

\item{metadata}{Service-specific metadata}

\item{labels}{Cross-service attributes for the location}
}
\value{
Location object
}
\description{
Location Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A resource that represents Google Cloud Platform location.
}
\seealso{
Other Location functions: \code{\link{Location.labels}},
  \code{\link{Location.metadata}}
}
