% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudkms_objects.R
\name{Rule}
\alias{Rule}
\title{Rule Object}
\usage{
Rule(notIn = NULL, description = NULL, conditions = NULL,
  logConfig = NULL, in. = NULL, permissions = NULL, action = NULL)
}
\arguments{
\item{notIn}{If one or more 'not_in' clauses are specified, the rule matches}

\item{description}{Human-readable description of the rule}

\item{conditions}{Additional restrictions that must be met}

\item{logConfig}{The config returned to callers of tech}

\item{in.}{If one or more 'in' clauses are specified, the rule matches if}

\item{permissions}{A permission is a string of form '<service>}

\item{action}{Required}
}
\value{
Rule object
}
\description{
Rule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A rule to be applied in a Policy.
}
