% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudkms_objects.R
\name{UpdateCryptoKeyPrimaryVersionRequest}
\alias{UpdateCryptoKeyPrimaryVersionRequest}
\title{UpdateCryptoKeyPrimaryVersionRequest Object}
\usage{
UpdateCryptoKeyPrimaryVersionRequest(cryptoKeyVersionId = NULL)
}
\arguments{
\item{cryptoKeyVersionId}{The id of the child CryptoKeyVersion to use as primary}
}
\value{
UpdateCryptoKeyPrimaryVersionRequest object
}
\description{
UpdateCryptoKeyPrimaryVersionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for KeyManagementService.UpdateCryptoKeyPrimaryVersion.
}
